refactor folder entities
