+++ src/admin/admin.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AdminController } from './admin.controller';
 import { AdminService } from './admin.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Admin } from './entities/admin.entity';
+import { Admin } from '../entities/admin.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Admin])],
