+++ src/app.module.ts
@@ -29,8 +29,10 @@ envConfig();
           username: process.env.DB_USERNAME,
           password: process.env.DB_PASSWORD,
           database: process.env.DB_NAME,
-          entities: [join(__dirname, '**/*.entity{.ts,.js}')],
+          // entities: [join(__dirname, '**/*.entity{.ts,.js}')],
+          entities: [join(__dirname, "entities/*.entity{.ts,.js}")],
           synchronize: true,
+          // logging: true,
         };
       },
     }),
