+++ src/category/category.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { CategoryController } from './category.controller';
 import { CategoryService } from './category.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Category } from './entities/category.entity';
+import { Category } from '../entities/category.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Category])],
