+++ src/product/product.module.ts
@@ -4,9 +4,9 @@ import { ProductService } from './product.service';
 // import { FirebaseModule } from 'firebase/firebase.module';
 import { FirebaseService } from 'firebase/firebaseService';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Category } from 'category/entities/category.entity';
-import { Product } from './entities/product.entity';
-import { SubCategory } from 'sub-category/entities/subCategory.entity';
+import { Category } from '../entities/category.entity';
+import { Product } from '../entities/product.entity';
+import { SubCategory } from '../entities/subCategory.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Category, Product, SubCategory])],
