+++ src/sub-category/sub-category.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { SubCategoryController } from './sub-category.controller';
 import { SubCategoryService } from './sub-category.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { SubCategory } from './entities/subCategory.entity';
-import { Category } from 'category/entities/category.entity';
+import { SubCategory } from '../entities/subCategory.entity';
+import { Category } from '../entities/category.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([SubCategory, Category])],
